+++ R/utils.R
@@ -106,8 +106,11 @@ download_to_temp <- function(url) {
   extt <- tools::file_ext(url)
   tmpfi <- tempfile(fileext = paste0(".", extt))
   rq <- httr2::request(url)
-  # nolint start
-  dwn <- httr2::req_perform(rq, path = tmpfi)
-  # nolint end
+  rq <- httr2::req_headers(
+    rq,
+    referer = "https://dieghernan.github.io/resmush/"
+  )
+
+  dwn <- httr2::req_perform(rq, path = tmpfi) # nolint
   tmpfi
 }